Move notification items to own module
Move the notification item to its own module, and remove the messages
module. This collects the notification-related behavior under one
module.
